+++ compiled/jquery-mobile-angular-adapter-standalone-1.1.2-SNAPSHOT.js
@@ -34176,9 +34176,6 @@ factory(window.jQuery, window.angular);
             var routeOverride = $location.$$routeOverride;
             delete $location.$$routeOverride;
             if (routeOverride) {
-                if (routeOverride.templateUrl) {
-                    newRoute.templateUrl = routeOverride.templateUrl;
-                }
                 if (routeOverride.onActivate) {
                     newRoute.onActivate = routeOverride.onActivate;
                 }
@@ -34202,7 +34199,7 @@ factory(window.jQuery, window.angular);
         function onPagebeforeshow(event) {
             var current = $route.current;
             if (current && current.onActivate) {
-                event.targetScope[current.onActivate].call(event.targetScope, current.locals);
+                event.targetScope.$eval(current.onActivate, current.locals);
             }
             var isDialog = $.mobile.activePage && $.mobile.activePage.jqmData("role") === "dialog";
             if (isDialog) {
@@ -34927,61 +34924,6 @@ factory(window.jQuery, window.angular);
     }
 
 })(angular);
-(function($, angular) {
-    function splitAtFirstColon(value) {
-        var pos = value.indexOf(':');
-        if (pos===-1) {
-            return [value];
-        }
-        return [
-            value.substring(0, pos),
-            value.substring(pos+1)
-        ];
-    }
-
-    var mod = angular.module('ng');
-    mod.factory('$navigate', ['$location', function($location) {
-        /*
-         * Service for page navigation.
-         * @param target has the syntax: [<transition>:]pageId
-         * @param activateFunctionName Function to call in the target scope.
-         * @param further params Parameters for the function that should be called in the target scope.
-         */
-        function navigate(target, activateFunctionName, activateParams) {
-            var navigateOptions;
-            if (typeof target === 'object') {
-                navigateOptions = target;
-                target = navigateOptions.target;
-            }
-            var parts = splitAtFirstColon(target);
-            var isBack = false;
-            if (parts.length === 2 && parts[0] === 'back') {
-                isBack = true;
-                target = parts[1];
-            } else if (parts.length === 2) {
-                navigateOptions = { transition: parts[0] };
-                target = parts[1];
-            }
-            if (target === 'back') {
-                $location.goBack();
-            } else {
-                if (isBack) {
-                    $location.backMode();
-                }
-                $location.url(target);
-            }
-            $location.routeOverride({
-                jqmOptions: navigateOptions,
-                onActivate: activateFunctionName,
-                locals: activateParams
-            });
-        }
-
-        return navigate;
-    }]);
-
-
-})($, angular);
 (function(angular) {
     var storageName = '$$sharedControllers';